+++ example/snippets/120-agent-initialization-dec112.spec.ts
@@ -11,16 +11,14 @@ it('Shows how to initialize an agent for a DEC112 environment', async () => {
     user: 'user',
     // DEC112 environments require a verified telephone number, due to legal requirements.
     displayName: '004366412345678',
-    namespaceSpecifics: new DEC112Specifics(
-      // device id (registration API version 1; deprecated)
-      undefined,
+    namespaceSpecifics: new DEC112Specifics({
       // registration id (registration API version 2)
-      'registrationId',
+      registrationId: 'registrationId',
       // user device language (ISO639-1 two letter language code; optional)
-      'en',
+      langauge: 'en',
       // client version (e.g. version of application, where ng112-js is used in; optional)
-      '1.0.4',
-    ),
+      clientVersion: '1.0.4',
+    }),
   });
 
   endExample();