+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This also means you can write your own adapter and use any SIP library of your c
 `JsSIP` and `SIP.js` are supported by default. Examples can be found under `./example/snippets/`.
 - **BREAKING**: New object for message send errors: `MessageError` instead of former object `MessageFailedEvent` \
 Object has different properties!
+- **BREAKING**: `DEC112Specifics` constructor uses object `DEC112Config` instead of individual parameters for initialization.
 ### Added
 - Agent mode that can be used to indicate an app is running in the background.
 - Support for DIDs.